backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
motor>=3.3.2
pymongo>=4.6.1
pydantic[email]>=2.7.4
pydantic-settings>=2.1.0
python-multipart>=0.0.6
langchain-google-genai>=2.0.0
langchain>=0.3.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt==4.0.1
email-validator>=2.0.0
